Working on adding an Area Code filter.
I think an enhancement might be adding the ability to only execute for lines of my Area. My company's olr model contains a LOT of outside Areas. I would prefer to not have to remove the other Areas just to run this script. Hard-coded to one Area now, but am working towards allowing either a range of areas (536-540), multiple unique areas (536, 338, 22), or some combination of the two.